use recoil in TeamsPage
